Fix export of `DrizzleQueryError` (fixes #4618)
There was a module name clash of `src/errors.ts` and
`src/errors/index.ts`. This commit moves `DrizzleQueryError` into
`src/errors.ts` and removes the module clash to fix the issue.
