+++ drizzle-orm/src/errors.ts
@@ -10,6 +10,20 @@ export class DrizzleError extends Error {
 	}
 }
 
+export class DrizzleQueryError extends Error {
+	constructor(
+		public query: string,
+		public params: any[],
+		public override cause?: Error,
+	) {
+		super(`Failed query: ${query}\nparams: ${params}`);
+		Error.captureStackTrace(this, DrizzleQueryError);
+
+		// ES2022+: preserves original error on `.cause`
+		if (cause) (this as any).cause = cause;
+	}
+}
+
 export class TransactionRollbackError extends DrizzleError {
 	static override readonly [entityKind]: string = 'TransactionRollbackError';
