--- drizzle-orm/src/errors/index.ts
@@ -1,13 +1,0 @@
-export class DrizzleQueryError extends Error {
-	constructor(
-		public query: string,
-		public params: any[],
-		public override cause?: Error,
-	) {
-		super(`Failed query: ${query}\nparams: ${params}`);
-		Error.captureStackTrace(this, DrizzleQueryError);
-
-		// ES2022+: preserves original error on `.cause`
-		if (cause) (this as any).cause = cause;
-	}
-}
